Atualizando main

+++ src/main/java/com/marmitech/Marmitech/DTO/RequestDTO/ProdutoSaveDTO.java
@@ -1,7 +1,4 @@
 package com.marmitech.Marmitech.DTO.RequestDTO;
 
-import java.math.BigDecimal;
-
-public record ProdutoSaveDTO(String nome, String descricao, String categoria, int estoque, BigDecimal precoUnitario, String sku) {
-
+public record ProdutoSaveDTO(String nome, String descricao, String categoria, int estoque, Double precoUnitario, String sku) {
 }
