+++ src/main/java/com/marmitech/Marmitech/DTO/ResponseDTO/ProdutoListaDTO.java
@@ -1,7 +1,5 @@
 package com.marmitech.Marmitech.DTO.ResponseDTO;
 
-import java.math.BigDecimal;
-
-public record ProdutoListaDTO(int id, String nome, String descricao, String categoria, String dataCadastro, BigDecimal precoUnitario, int estoque, String sku) {
+public record ProdutoListaDTO(int id, String nome, String descricao, String categoria, String dataCadastro, Double precoUnitario, int estoque, String sku) {
 
 }
